Updated readme intro paragraph, moved app examples into seperate file.
